requirements.txt
attrs==17.4.0
certifi==2018.1.18
chardet==3.0.4
funcsigs==1.0.2
graphene==2.0.1
graphql-core==2.0
graphql-relay==0.4.5
graphql-server-core==1.0.dev20170322001
idna==2.6
Paste==2.0.3
pkginfo==1.4.1
pluggy==0.6.0
promise==2.1
py==1.5.2
requests==2.18.4
requests-toolbelt==0.8.0
Rx==1.6.1
six==1.11.0
tqdm==4.19.6
twine==1.9.1
typing==3.6.4
urllib3==1.22
web.py==0.39
Werkzeug==0.14.1
